+++ next/portal_cms_demo_next/src/config/data.ts
@@ -1,3 +1,4 @@
+import {ChangeCounselData, CounselData, ListData, MonthAccidentData, UserType} from "@/@types/common";
 
 
 export const initialCounselData: CounselData[] = [
@@ -51,25 +52,8 @@ export const initialCounselData: CounselData[] = [
     },
 ];
 
+
 export const changeCounselData: ChangeCounselData[] = [
-    {
-        cNo: 8,
-        cDay: "2023-08",
-        pNo: "FA20235204423000",
-        pAdd: 12,
-        pEnd: 4,
-        AddAmt: 168178432,
-        EndAmt: 1917524,
-    },
-    {
-        cNo: 7,
-        cDay: "2023-07",
-        pNo: "FA20235204423000",
-        pAdd: 3,
-        pEnd: 2,
-        AddAmt: 854354,
-        EndAmt: 638544,
-    },
     {
         cNo: 6,
         cDay: "2023-06",
@@ -152,18 +136,6 @@ export const monthAccidentData: MonthAccidentData[] = [
         suspense: 4,
     },
 ]
-
-export const topCounselData = {
-    labels: ['정곡빌딩', '부산 사학연금', '청주성모병원', '서울스퀘어', '일산국립암센터'],
-    values: [2535000, 2282000, 1650000, 1609000, 1150000],
-    color: '#4a7ff7'
-};
-
-export const topBusinessData = {
-    labels: ['F1963 1 주차장', '가든호텔', '그랜드하얏인천', '그레이츠판교', '명지병원'],
-    values: [4, 3, 3, 2, 1],
-    color: '#4a7ff7'
-};
 
 export const listData: ListData[] = [
     {
