+++ next/portal_cms_demo_next/src/config/themeConfig.ts
@@ -1,10 +1,7 @@
 import HiparkingLogo from '@/assets/images/logo/hiparking-logo.png'
 import SimgLogo from '@/assets/images/logo/simg-round-logo.png'
-import KmparkLogo from '@/assets/images/logo/kmpark-logo.svg'
-import {changeCounselData} from "@/config/data";
+import KmparkLogo from '@/assets/images/logo/kmpark-logo.png'
 import {StaticImageData} from "next/image";
-import {TooltipItem } from 'chart.js';
-import {Context} from "chartjs-plugin-datalabels";
 
 
 interface Theme {
@@ -39,7 +36,7 @@ const themeConfigs: ThemeConfig = {
     kmpark: {
         logoSrc: KmparkLogo,
         menuItems: {
-            dashboard: "/hiparking",
+            dashboard: "/kmpark",
             list: "/kmpark/list",
             mypage: "/kmpark/mypage",
         },
@@ -52,142 +49,3 @@ export const getThemeConfig = (theme: string): Theme => {
 };
 
 export const availableThemes = Object.keys(themeConfigs);
-
-//그래프옵션
-export const optionHiparkingTwowayBar = {
-    responsive: true,
-    scales: {
-        x: {
-            stacked: true
-        },
-        y: {
-            beginAtZero: true,
-            ticks: {
-                stepSize: 5,
-            },
-            grid: {
-                display: false,
-            },
-        },
-    },
-    plugins: {
-        legend: {
-            display: true,
-        },
-        tooltip: {
-            backgroundColor: 'white',
-            titleColor: 'black',
-            bodyColor: 'black',
-            borderWidth: 1,
-            borderColor: '#e7e7e7',
-            bodyAlign: 'center' as const,
-            titleAlign: 'center' as const,
-            position: 'average' as const,
-            yAlign: 'bottom' as const,
-            callbacks: {
-                label: (context: TooltipItem<'bar'>) => {
-                    const dataIndex = context.dataIndex;
-                    const datasetIndex = context.datasetIndex;
-                    if (datasetIndex === 0) {
-                        return [
-                            `추가 사업장: ${changeCounselData[dataIndex].pAdd}`,
-                            `추가 보험금: ${changeCounselData[dataIndex].AddAmt.toLocaleString()} 원`,
-                        ];
-                    } else {
-                        return [
-                            `종료 사업장: ${changeCounselData[dataIndex].pEnd}`,
-                            `감소 보험금: ${changeCounselData[dataIndex].EndAmt.toLocaleString()} 원`,
-                        ];
-                    }
-                },
-            },
-        },
-    },
-};
-
-export const optionHiparkingBarHorizon = {
-    indexAxis: 'y' as const,
-    responsive: true,
-    plugins: {
-        legend: {
-            display: false,
-        },
-        tooltip: {
-            enabled: false,
-        },
-    },
-    scales: {
-        x: {
-            beginAtZero: true,
-            grid: {
-                display: false,
-            },
-            ticks: {
-                display: false,
-            },
-        },
-        y: {
-            grid: {
-                display: false,
-            },
-            ticks: {
-                font: {
-                    size: 15,
-                },
-            },
-        },
-    },
-    layout: {
-        padding: {
-            right: 60,
-        },
-    },
-    cutout: '75%',
-};
-
-
-export const optionHiparkingDoughnut = {
-    responsive: true,
-    plugins: {
-        tooltip: {
-            enabled: false,
-        },
-    },
-    cutout: '75%',
-};
-
-export const optionHiparkingPie = {
-    plugins: {
-        legend: {
-            display: false,
-        },
-        tooltip: {
-            backgroundColor: 'white',
-            titleColor: 'black',
-            bodyColor: 'black',
-            borderWidth: 1,
-            borderColor: '#e7e7e7',
-            bodyAlign: 'center',
-            titleAlign: 'center',
-            position: 'nearest',
-            yAlign: 'bottom',
-        },
-        datalabels: {
-            formatter: function (value: number, context: Context) {
-                const dataset = context.chart.data.datasets[0];
-                const total = dataset.data.reduce((acc: number, val: unknown) => acc + (typeof val === 'number' ? val : 0), 0);
-                if (total === 0) return '0%';
-                const percentage = ((value / total) * 100).toFixed(0) + "%";
-                return percentage;
-            },
-            color: '#fff',
-            anchor: 'center',
-            align: 'center',
-            font: {
-                size: 15,
-                weight: 'normal',
-            },
-        },
-    },
-    responsive: false,
-};
