+++ next/portal_cms_demo_next/src/styles/globals.css
@@ -112,8 +112,8 @@ input[type=file]::file-selector-button {
         --color-sub-lighter: #fadfaf;
     }
     html[data-theme='kmpark'] {
-        --color-main: #4a80f7;
-        --color-main-light: #79a2fc;
+        --color-main: #6a95f6;
+        --color-main-light: #a3bbef;
         --color-main-lighter: #e0e9ff;
         --color-sub: #868686;
         --color-sub-light: #c0c0c0;
